backend/requirements.txt
asgiref==3.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==3.0.5
django-cors-headers==3.11.0
djangorestframework==3.11.0
idna==3.3
ndg-httpsclient==0.5.1
psycopg2==2.8.6
pusher==3.2.0
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9